move class def
